Introducing header file for ippusbxd and adding documentation
This change moves the struct and function declarations which located
at the beginning of ippusbxd.c into their own dedicated header and
adds some more documentation. This should help to simplify the code
somewhat and make it easier to look up the expected behavior of a
function.
